refactor: 리뷰 반영 및 주석, log 지우기
