+++ be/algo-with-me-score/src/app.module.ts
@@ -1,17 +1,20 @@
 import { BullModule } from '@nestjs/bull';
 import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import * as dotenv from 'dotenv';
 
 import * as process from 'node:process';
 
-import { Submission } from './entities/submission.entity';
-import { SubmissionConsumer } from './services/score.consumer';
-
-dotenv.config();
+import { Problem } from './score/entities/problem.entity';
+import { Submission } from './score/entities/submission.entity';
+import { ScoreModule } from './score/score.module';
 
 @Module({
   imports: [
+    ConfigModule.forRoot({
+      cache: true,
+      isGlobal: true,
+    }),
     TypeOrmModule.forRoot({
       type: 'postgres',
       host: process.env.DB_HOST,
@@ -19,10 +22,9 @@ dotenv.config();
       username: process.env.DB_USERNAME,
       password: process.env.DB_PASSWORD,
       database: process.env.DB_NAME,
-      synchronize: true,
-      entities: [],
+      synchronize: false,
+      entities: [Submission, Problem],
     }),
-    TypeOrmModule.forFeature([Submission]),
     BullModule.forRoot({
       redis: {
         host: process.env.REDIS_HOST,
@@ -30,11 +32,9 @@ dotenv.config();
         password: process.env.REDIS_PASSWORD,
       },
     }),
-    BullModule.registerQueue({
-      name: 'testQueue',
-    }),
+    ScoreModule,
   ],
   controllers: [],
-  providers: [SubmissionConsumer],
+  providers: [],
 })
 export class AppModule {}
