+++ be/algo-with-me-score/src/score/entities/submission.enums.ts
