+++ be/algo-with-me-score/src/score/interfaces/coderun-response.interface.ts
