+++ be/algo-with-me-score/src/score/interfaces/score-result.interface.ts
